backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
openai==0.27.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21